% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gisfao.R
\docType{SpatialPointsDataFrame}
\name{fao_world_centroids}
\alias{fao_world_centroids}
\title{fao_world_centroids SpatialPointsDataFrame}
\format{A SpatialPointsDataFrame with 187 rows and 19 variables}
\source{
\url{http://www.fao.org/geonetwork/srv/en/metadata.show?id=12691}
}
\usage{
data(fao_world_centroids)
}
\description{
A SpatialPointsDataFrame containing the polygon centroids all the countries in the world
accompanyig with country coding compatible with UNFAO FAOST_CODE coding.
}
\details{
Variables in \code{data} slot are

\itemize{
 \item OBJECTID
 \item STATUS
 \item STR_YEAR
 \item EXP_YEAR
 \item DISP_AREA
 \item ADM0_CODE
 \item ADM0_NAME Short name
 \item UN_CODE United Nations country code
 \item FAO_CODE Country code used by UNFAO, identical with FAOST_CODE
 \item Shape_Leng
 \item Shape_Area
 \item MaxSimpTol
 \item MinSimpTol
 \item FAOST_CODE Country code used by UNFAO, identical with FAOST_CODE
 \item RAF Logical whether county belongs to Africa in UNFAO regional classification
 \item LAC Logical whether county belongs to Latin America and the Caribbean in UNFAO regional classification
 \item RAP Logical whether county belongs to Asia and the Pacific in UNFAO regional classification
 \item REU Logical whether county belongs to Europe and Central Asia in UNFAO regional classification
 \item RNE Logical whether county belongs to Near East and North Africa in UNFAO regional classification
}
}
\references{
\url{http://www.fao.org/geonetwork/srv/en/metadata.show?id=12691}
}
\keyword{shapefile}

