% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/load_fao_shape.R
\name{load_fao_shape}
\alias{load_fao_shape}
\title{Load FAOSTAT  compatible shapefile}
\usage{
load_fao_shape(region = "world")
}
\arguments{
\item{region}{A string "RAF", "LAC", "RAP", "REU", "RNE". Default is "world"}
}
\value{
a SpatialPolygonDataFrame
}
\description{
Load FAOSTAT  compatible shapefile
}
\examples{
africa <- load_fao_shape(region = "RAF")
plot(africa)
}

